move helper functions from overview into data_utils
